railway/requirements.txt
mysql-connector-python
gradio
pandas
